Update translation.json
I changed all the in-game names to the ones used by Hello Games. So if they seem odd, it's because they are and it's HG's fault. :) 

Other than that, I made some minor changes, such as replacing a word here and there, and replacing all formal addressing with a regular one for consistency.